requirements.txt
tqdm==4.65.0
pyyaml==6.0
matplotlib==3.7.2
numpy==1.25.1